Many Bugfixes, made the install routine asynchronous so the tasks are waiting correctly
Those changes are made to avoid many "file-not-found"-Exceptions.

Corrected the check for the EAX-Support in Options-Window.

Added Manifest for Updater-Binary to be able to launch the main BFME1-Form as admin as well.

Brutal-AI file was not shipped properly -> fixed.

Added logic to Install-Path-DialogResult so we dont start install routine, when User presses "X".

Added Registry StartMenu-Key-Reading for later feature.
Changed method-name from "RandomString" to "CDKeyGenerator" to make it more obvious.

The Install-Routine is now much more error-resistant.