Launcher Release 1.0.1.14
Fixed an Issue, where the Launcher is always installing a Patch after each start.
Fixed an Issue, where the Launcher is installing Patch 3.0 after installing the game instead of Patch 3.1.
Fixed an Issue, where the Launcher was not able to retrieve the newest Beta files.
Fixed an Issue, where the Launcher destroyed a Beta-File while downloading.